+++ Tests/WindowsCloudPC.Tests.ps1
@@ -3,8 +3,7 @@ BeforeDiscovery {
         'Connect-CloudPC',
         'Get-CloudPC',
         'Get-CloudPCProvisioningPolicy',
-        'Get-CloudPCUsage',
-        'Get-CloudPCUsageBeta'
+        'Get-CloudPCUsage'
     )
 }